dependencies and colours

+++ build.sh
@@ -1,28 +1,36 @@
 #!/bin/bash
+red='\033[0;31m'
+blue='\033[1;34m'
+yellow='\033[1;33m'
+NC='\033[0m'
+if [[ $EUID -ne 0 ]]; then
+   echo -e "${red}This script must be run as root${NC}" 1>&2
+   exit 1
+fi
 build="../build/panda"
 cd ./src
 fcount=`ls -l ../build/old-builds | wc -l`
 if [ $fcount -ge 5 ]; then
-	echo "Erasing old builds"
+	echo -e "${blue}Erasing old builds${NC}"
 	cd ../build/old-builds
 	rm -R panda*
 	cd ../../src
 fi
 if [ -f $build ]; then
-	echo "Moving old build"
+	echo -e "${blue}Moving old build${NC}"
 	mv $build "../build/old-builds/panda-`date +%s`"
 fi
-echo "Starting GHC"
+echo -e "${blue}Starting GHC${NC}"
 /usr/bin/time -f "%S" ghc --make -o $build Main 2> temp
 ptime=`cat temp`
 rm temp
 cd ..
 if [ -f "panda" ]; then
-	echo "Deleting old symbolic link"
+	echo -e "${blue}Deleting old symbolic link${NC}"
 	rm panda
 fi
 if [ -f "build/panda" ]; then
-	echo "Creating symbolic link"
+	echo -e "${blue}Creating symbolic link${NC}"
 	ln -s build/panda panda
 fi
-echo "Done in "$ptime"s"
+echo -e "${blue}Done in ${yellow}"$ptime"s${NC}"
